+++ app/controllers/cart_items_controller.rb
@@ -1,11 +1,10 @@
 class CartItemsController < ApplicationController
-  before_action :set_cat, only: [:create, :delete]
 
   def show
     cats = []
     quantity = []
-    if params[:cart_contents]
-    params[:cart_contents].each do |cats_and_quantity|
+    if session[:cart_data]
+    session[:cart_data].each do |cats_and_quantity|
       quantity << cats_and_quantity[1].to_i
       cats << Cat.find(cats_and_quantity[0])
     end
@@ -19,21 +18,15 @@ class CartItemsController < ApplicationController
   end
 
   def create
-    @cart.add_to_cart(@cat.id)
+    @cart.add_to_cart(params[:id])
     session[:cart_data] = @cart.contents
-    redirect_to cart_path(cart_contents: @cart.contents), alert: "cat added to cart"
+    redirect_to cart_path, alert: "cat added to cart"
   end
 
   def delete
-    @cart.remove_from_cart(@cat.id)
+    @cart.remove_from_cart(params[:id])
     session[:cart_data] = @cart.contents
     redirect_to cart_path(cart_contents: @cart.contents)
-  end
-
-  def set_cat
-    if params[:id]
-      @cat = Cat.find(params[:id])
-    end
   end
 
   def total_price(c_and_q)
