+++ app/controllers/order_cats_controller.rb
@@ -1,7 +1,7 @@
 class OrderCatsController < ApplicationController
-
+  include ResetCart
   def create
-    authorize! :create, @order
+    authorize! :create, Order
     user = current_user
     if session[:checkout_order]
       order = user.orders.create
@@ -14,14 +14,4 @@ class OrderCatsController < ApplicationController
       redirect_to cart_path
     end
   end
-
-  private
-
-  def reset_cart
-    session[:checkout_order] = nil
-    session[:cart_items] = nil
-    session[:cart_data] = nil
-    session[:cart_contents] = nil
-  end
-
 end
